Remove charisma game mode, convert the variables from it into a UObject. Refactor BP game mode and meta functions to accomodate
