+++ Source/CharismaPlugNPlay/Public/Playthrough/CharismaConnectionParams.h
@@ -1,0 +1,22 @@
+﻿#pragma once
+
+#include "CoreMinimal.h"
+#include "CharismaConnectionParams.generated.h"
+
+USTRUCT(BlueprintType)
+struct CHARISMAPLUGNPLAY_API FCharismaConnectionParams
+{
+	GENERATED_USTRUCT_BODY()
+
+	UPROPERTY(BlueprintReadWrite, EditAnywhere, Category = "Charisma|GameMode")
+	int StoryId;
+
+	UPROPERTY(BlueprintReadWrite, EditAnywhere, Category = "Charisma|GameMode")
+	int StoryVersion;
+	
+	UPROPERTY(BlueprintReadWrite, EditAnywhere, Category = "Charisma|GameMode")
+	FString ApiKey;
+
+	UPROPERTY(BlueprintReadWrite, EditAnywhere, Category = "Charisma|GameMode")
+	FString StartGraphRefId;
+};
